Organization & hidden object instantiation
Drastically changed directory structure for better organization. Removed testing results as we are not really ready for testing yet. I want to finish getting the items instantiated into the UI Documents with the right position, size, and rotation. For now, I am happy that we can get to the point of loading the level with the items actually showing up.